+++ packages/basehub/src/bin/util/get-stuff-from-env.ts
@@ -80,11 +80,15 @@ export const getStuffFromEnv = (
     parsedDebugForcedURL ? parsedDebugForcedURL : `${basehubAPIOrigin}/graphql`
   );
 
+  let tokenNotFoundErrorMessage = `🔴 Token not found. Make sure to include the ${buildEnvVarName(
+    "TOKEN"
+  )} env var.`;
+
   const resolveTokenParam = (token: string | null) => {
     if (!token) return null;
-    console.warn(
-      `🚧 Warning! The --token parameter will be deprecated in the next major version. You should use --env-prefix instead.`
-    );
+    // console.warn(
+    //   `🚧 Warning! The --token parameter will be deprecated in the next major version. You should use --env-prefix instead.`
+    // );
     const isRaw = token.startsWith("bshb_");
     if (isRaw) {
       console.warn(
@@ -92,6 +96,7 @@ export const getStuffFromEnv = (
       );
       return token;
     }
+    tokenNotFoundErrorMessage = `🔴 Token not found. Make sure to include the ${token} env var.`;
     const fromEnv = process.env[token];
     if (fromEnv) return fromEnv;
     return ""; // empty string to prevent fallback
@@ -109,11 +114,7 @@ export const getStuffFromEnv = (
     null;
 
   if (!token) {
-    console.log(
-      `🔴 Token not found. Make sure to include the ${buildEnvVarName(
-        "TOKEN"
-      )} env var.`
-    );
+    console.log(tokenNotFoundErrorMessage);
     process.exit(1);
   }
 
@@ -254,10 +255,15 @@ export const getStuffFromEnv = (options) => {
     const parsedBasehubDraftEnv = getEnvVar("DRAFT");
     const parsedBasehubApiVersionEnv = getEnvVar("API_VERSION");
 
+    let tokenNotFoundErrorMessage = \`🔴 Token not found. Make sure to include the \${buildEnvVarName(
+      "TOKEN"
+    )} env var.\`;
+
     const resolveTokenParam = (token) => {
       if (!token) return null;
       const isRaw = token.startsWith("bshb_");
       if (isRaw) return token;
+      tokenNotFoundErrorMessage = \`🔴 Token not found. Make sure to include the \${token} env var.\`;
       return process.env[token] ?? ''; // empty string to prevent fallback
     };
 
@@ -272,11 +278,7 @@ export const getStuffFromEnv = (options) => {
       null;
 
     if (!token) {
-      throw new Error(
-        \`🔴 Token not found. Make sure to include the \${
-          buildEnvVarName("TOKEN")
-        } env var.\`
-      );
+      throw new Error(tokenNotFoundErrorMessage);
     }
 
     const ref =